requirements.txt
rich~=13.7.1
typer~=0.12.3
uvicorn~=0.30.1
numpy~=1.26.4
jieba~=0.42.1
colorlog~=6.8.2
requests~=2.32.3
huggingface-hub~=0.23.4
filelock~=3.15.4
tqdm~=4.66.4
pillow~=10.4.0
packaging~=24.1
pip~=24.0
safetensors~=0.4.3
datasets~=2.20.0
fastapi~=0.111.0
pydantic~=2.8.2
scikit-learn~=1.5.1
seqeval~=1.2.2
pandas~=2.2.2
multiprocess~=0.70.12.2
scipy~=1.14.0
six~=1.16.0
sentencepiece~=0.2.0
regex~=2024.5.15
Jinja2~=3.1.4
holidays~=0.52
openai~=1.35.13
paddlepaddle-gpu>=2.3.0
